requirements.txt
numpy
streamlit
streamlit-webrtc
pillow
mediapipe
opencv-python